feat: デッキカードとフォーム入力のハンドラーを統合し、モーダル操作を改善
